Add some coroutine helpers; simplify error handling in sync code
